requirements.txt
langchain 
langchain-community 
langchain-huggingface 
langchain-groq 
chromadb 
pypdf 
python-dotenv
streamlit
sentence-transformers